TRestHitsGaussAnalysisProcess generalization of readout parameters
When the process was added in commit 4c60811d96ffb8b74007aaa9b3c9da4432f5f590 the histogram
binning was hardcoded to match the IAXO-D0 readout (number of strips and pitch). Now these
parameters can be defined by the user in the RML file, being the default values those
of IAXO-D0.