fix `gDirectory` comparison for ROOT 6.24 / C++14
To be honest I don't know which of the two is the cause. My local ROOT
installation of version 6.24 (release candidate 1) doesn't compile
unless I compile with C++14. Doing so still leads to an issue here
where the compiler suddenly cannot resolve which overload to choose
for the `!=` operator.

Stupid implicit conversions...