[unix] do not use $ROOTSYS, use `root-config --bindir`
Arguably one might want to error out same as for windows here. But
given that `root-config` is enough to identify the bin directory of
root, in theory it's enough for our purposes.

Without the definition of `ROOTSYS` (i.e. `thisroot` wasn't sourced),
hard to debug errors appear for people unfamiliar with ROOT & CMAKE